chore: extract test harness and OTel plumbing into src/runner.ts
Part of REAL-1037.
